# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UrbanThreads e-commerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProductGender = typeof ProductGender[keyof typeof ProductGender];


export const ProductGender = {
  mens: 'mens',
  womens: 'womens',
  unisex: 'unisex',
} as const;

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  salePrice?: number | null;
  gender: ProductGender;
  category: string;
  sizes: string[];
  imageUrl: string;
  imageUrls?: string[];
  inStock: boolean;
  featured?: boolean;
  trending?: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount?: number;
}

export type CategoryGender = typeof CategoryGender[keyof typeof CategoryGender];


export const CategoryGender = {
  mens: 'mens',
  womens: 'womens',
  unisex: 'unisex',
} as const;

export interface Category {
  id: number;
  name: string;
  gender: CategoryGender;
  slug: string;
  productCount?: number;
}

export interface StorefrontSummary {
  totalProducts: number;
  mensProducts: number;
  womensProducts: number;
  featuredProducts: Product[];
  trendingProducts?: Product[];
  categories?: Category[];
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productImageUrl: string;
  price: number;
  size: string;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  itemCount: number;
}

export interface CartItemInput {
  productId: number;
  size: string;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface OrderItem {
  id: number;
  productId: number;
  productName: string;
  productImageUrl: string;
  price: number;
  size: string;
  quantity: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  pending: 'pending',
  confirmed: 'confirmed',
  processing: 'processing',
  shipped: 'shipped',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface Address {
  id: number;
  fullName: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  phone: string;
  isDefault?: boolean;
}

export interface Order {
  id: number;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  shippingAddress?: Address;
  /** @nullable */
  trackingNumber?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
}

export interface OrderInput {
  addressId: number;
  paymentMethod: string;
}

export interface AddressInput {
  fullName: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  phone: string;
  isDefault?: boolean;
}

export type ListProductsParams = {
category?: string;
gender?: ListProductsGender;
search?: string;
minPrice?: number;
maxPrice?: number;
featured?: boolean;
limit?: number;
offset?: number;
};

export type ListProductsGender = typeof ListProductsGender[keyof typeof ListProductsGender];


export const ListProductsGender = {
  mens: 'mens',
  womens: 'womens',
  unisex: 'unisex',
} as const;