# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: UrbanThreads e-commerce API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Product catalog operations
  - name: cart
    description: Shopping cart operations
  - name: orders
    description: Order management
  - name: addresses
    description: Delivery address management
  - name: categories
    description: Category operations
  - name: storefront
    description: Storefront summary and featured content

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: gender
          in: query
          schema:
            type: string
            enum: [mens, womens, unisex]
        - name: search
          in: query
          schema:
            type: string
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured products for homepage hero
      responses:
        "200":
          description: Featured products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/trending:
    get:
      operationId: getTrendingProducts
      tags: [products]
      summary: Get trending/best-selling products
      responses:
        "200":
          description: Trending products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List all product categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /storefront/summary:
    get:
      operationId: getStorefrontSummary
      tags: [storefront]
      summary: Get homepage storefront summary (counts, hero data)
      responses:
        "200":
          description: Storefront summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorefrontSummary"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get the current user's cart
      responses:
        "200":
          description: Current cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: clearCart
      tags: [cart]
      summary: Clear all items from the cart
      responses:
        "200":
          description: Cart cleared

  /cart/items:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add an item to the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "201":
          description: Item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/items/{itemId}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update quantity of a cart item
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemUpdate"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove an item from the cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: Get the current user's order history
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place an order from the current cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a single order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Not found

  /addresses:
    get:
      operationId: listAddresses
      tags: [addresses]
      summary: List saved delivery addresses
      responses:
        "200":
          description: List of addresses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Address"
    post:
      operationId: createAddress
      tags: [addresses]
      summary: Add a new delivery address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddressInput"
      responses:
        "201":
          description: Address created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Address"

  /addresses/{id}:
    delete:
      operationId: deleteAddress
      tags: [addresses]
      summary: Delete a saved address
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Address deleted

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, name, price, gender, category, sizes, imageUrl, inStock]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        salePrice:
          type: ["number", "null"]
        gender:
          type: string
          enum: [mens, womens, unisex]
        category:
          type: string
        sizes:
          type: array
          items:
            type: string
        imageUrl:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        inStock:
          type: boolean
        featured:
          type: boolean
        trending:
          type: boolean
        rating:
          type: ["number", "null"]
        reviewCount:
          type: integer

    Category:
      type: object
      required: [id, name, gender, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        gender:
          type: string
          enum: [mens, womens, unisex]
        slug:
          type: string
        productCount:
          type: integer

    StorefrontSummary:
      type: object
      required: [totalProducts, mensProducts, womensProducts, featuredProducts]
      properties:
        totalProducts:
          type: integer
        mensProducts:
          type: integer
        womensProducts:
          type: integer
        featuredProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        trendingProducts:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        categories:
          type: array
          items:
            $ref: "#/components/schemas/Category"

    CartItem:
      type: object
      required: [id, productId, productName, productImageUrl, price, size, quantity]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
        price:
          type: number
        size:
          type: string
        quantity:
          type: integer

    Cart:
      type: object
      required: [items, subtotal, itemCount]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        itemCount:
          type: integer

    CartItemInput:
      type: object
      required: [productId, size, quantity]
      properties:
        productId:
          type: integer
        size:
          type: string
        quantity:
          type: integer

    CartItemUpdate:
      type: object
      required: [quantity]
      properties:
        quantity:
          type: integer

    OrderItem:
      type: object
      required: [id, productId, productName, productImageUrl, price, size, quantity]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productImageUrl:
          type: string
        price:
          type: number
        size:
          type: string
        quantity:
          type: integer

    Order:
      type: object
      required: [id, status, items, total, createdAt]
      properties:
        id:
          type: integer
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          type: number
        shippingAddress:
          $ref: "#/components/schemas/Address"
        trackingNumber:
          type: ["string", "null"]
        estimatedDelivery:
          type: ["string", "null"]
        createdAt:
          type: string

    OrderInput:
      type: object
      required: [addressId, paymentMethod]
      properties:
        addressId:
          type: integer
        paymentMethod:
          type: string

    Address:
      type: object
      required: [id, fullName, line1, city, state, postalCode, country, phone]
      properties:
        id:
          type: integer
        fullName:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        phone:
          type: string
        isDefault:
          type: boolean

    AddressInput:
      type: object
      required: [fullName, line1, city, state, postalCode, country, phone]
      properties:
        fullName:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        postalCode:
          type: string
        country:
          type: string
        phone:
          type: string
        isDefault:
          type: boolean