# lib/api-zod/src/generated/types/addressInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UrbanThreads e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface AddressInput {
  fullName: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  phone: string;
  isDefault?: boolean;
}