# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UrbanThreads e-commerce API
 * OpenAPI spec version: 0.1.0
 */
import type { Address } from './address';
import type { OrderItem } from './orderItem';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  status: OrderStatus;
  items: OrderItem[];
  total: number;
  shippingAddress?: Address;
  /** @nullable */
  trackingNumber?: string | null;
  /** @nullable */
  estimatedDelivery?: string | null;
  createdAt: string;
}