# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * UrbanThreads e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface OrderInput {
  addressId: number;
  paymentMethod: string;
}